test: move failover arg-construction tests into test/unit/commands
test_failover_abort_sends_abort_only, test_failover_force_without_target_omits_force,
and test_failover_builds_full_arg_list_in_order stub send_command and never
touch a server, so they belong in test/unit rather than mixed into the
standalone-only failover integration module.

capture_failover_args previously stubbed send_command on `r`, the live
TCP-connected client from Helper::Generic#setup. Replaced with an
unconnected_client fixture that stubs the FFI connect call, matching the
technique already used in generic_commands_test.rb and
connection_config_test.rb, so the suite runs with no server.
